requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
comm==0.2.2
datasets==2.20.0
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
donut-python==1.0.9
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
-e git+https://github.com/elderprince/HeR-T-Fine-tuning.git@65eda7148fb4f6d71779e2f2e940c4443fa7191d#egg=her_t_pytorch_lightning
huggingface-hub==0.24.5
idna==3.4
importlib_metadata==8.2.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
lightning-utilities==0.11.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==3.20.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
Pygments==2.18.0
pytest==8.3.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.2
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
regex==2024.7.24
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.4
sconf==0.2.5
sentencepiece==0.2.0
sentry-sdk==2.12.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
timm==1.0.8
tokenizers==0.13.3
tomli==2.2.1
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchmetrics==1.4.1
torchvision==0.16.0+cu121
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.25.1
triton==2.1.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.13
wandb==0.16.5
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
zss==1.2.0
